New EnchantSpell Targeting and Features
**Example Config**
EnchantSpell:
    spell-class: ".instant.EnchantSpell"
    enchantments:
      DAMAGE_ALL: 5
      PROTECTION_ENVIRONMENTAL: 1
    cast-item: 276
    inventory-type: "inventory"
    safe-enchantment: true
    random-enchantment-level: true
    force-level: false
    toggle: true

This spell **safely** enchants all items within the player's inventory with sharpness 1-5 or protection 1 wherever applicable. If any of these enchantments are on the items, it will **toggle** them off.

inventory-type: choose between mainHand, offHand, wearing and inventory. By default it is mainHand.

safe-enchantment: this will check whether or not it can apply the enchant and if the enchantment conflicts with another.

random-enchant-level: randomizes the level will be assigned to the enchantment where the minimum is 1 and the maximum is the value put for the enchantment.

force-level: This is mutually exclusive with toggle, this will force the level of the enchantment irrespective of the current enchantment level on the item.

toggle: this is mutually exclusive with force-level. This will remove the enchantment if it is on the item. If casted again, it will re-apply the enchantment.