Rotate Spell | affectPitch, Face and Mimic Options
Example Rotate Spell Config

RotateSpell:
  spell-class: ".targeted.RotateSpell"
  rotation: 10
  rotation-pitch: 0
  affect-pitch: false
  mimic-direction: false
  face-target: false
  face-caster: false

Notes: rotation only affects the Yaw value of an entity

New Features:

rotation-pitch: adds/remove pitch of an entity

affect-pitch: a safe flag for if you want the pitch unaffected. This must be true for rotation-pitch to work

mimic-direction: mimics the direction of the target/caster so that the target/caster faces in the same relative direction.

face-caster: forces the target to face the caster

face-target: forces the caster to face the target

Cool Ideas:

You can make a cut scene with the new rotate spell features. Simple set face-target to true and have a mob be followed as it moves across a screen.

You can make a dummy npc that faces your direction by constantly spamming it with face-caster rotate spells.

Sky's the limit...