Leap Spell Targeting Features
allow-targeting is false by default to support legacy spells.
This must be true to use the new leap spell targeting features

You can now cast a leap spell and have the caster leap in the direction of the target location or entity. You must invoke it within a TargetedMultiSpell to achieve this effect.