Force, Leap, Velocity spells can add velocity instead of set.
Spells affected:

Forcepush, ForceToss, ForceBomb, Leap, Velocity.

All these spells have an add-velocity-instead option that adds that force to the target instead of setting it.

If I had 3 velocity to the right and I add 2 to the left. I will move slower.

Perfect for black-hole or gravity spells. This option is false by default.